Rename the constants.js file in js/src/components/paid-ads to asset-specs.js
Address https://github.com/woocommerce/google-listings-and-ads/pull/1863#discussion_r1087576526
