+++ js/src/components/paid-ads/asset-specs.js
